pyproject.toml
[tool.poetry]
name = "poetry new myawesome"
version = "0.1.0"
description = "demo web"
authors = ["ouyangchenling"]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.16.0"
SQLAlchemy = "^1.4.29"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
fastapi = "^0.71.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
